Migrate SDF loader to spawn scenes
