+++ rmf_site_editor/src/site/georeference.rs
@@ -300,7 +300,7 @@ fn spawn_tile(
     };
     let quad_handle = meshes.add(mesh);
 
-    let texture_handle: Handle<Image> = asset_server.load(PathBuf::from(tile));
+    let texture_handle: Handle<Image> = asset_server.load(PathBuf::from(tile.clone()));
     let material_handle = materials.add(StandardMaterial {
         base_color_texture: Some(texture_handle.clone()),
         alpha_mode: AlphaMode::Blend,
