+++ rmf_site_editor/src/site/mod.rs
@@ -188,6 +188,16 @@ impl Plugin for SitePlugin {
         .init_resource::<PhysicalLightToggle>()
         .init_resource::<UpdateFuelCacheChannels>()
         .init_resource::<ModelTrashcan>()
+        .register_type::<NameInSite>()
+        .register_type::<AssetSource>()
+        .register_type::<Pose>()
+        .register_type::<IsStatic>()
+        .register_type::<Scale>()
+        .register_type::<ModelMarker>()
+        .register_type::<VisualMeshMarker>()
+        .register_type::<CollisionMeshMarker>()
+        .register_type::<Category>()
+        .register_type::<PrimitiveShape>()
         .add_event::<LoadSite>()
         .add_event::<ImportNavGraphs>()
         .add_event::<ChangeCurrentSite>()
@@ -385,7 +395,6 @@ impl Plugin for SitePlugin {
                 update_model_scenes,
                 update_affiliations,
                 update_members_of_groups.after(update_affiliations),
-                handle_new_sdf_roots,
                 update_model_scales,
                 make_models_selectable,
                 propagate_model_render_layers,
