+++ rmf_site_format/src/category.rs
@@ -16,7 +16,10 @@
 */
 
 #[cfg(feature = "bevy")]
-use bevy::prelude::Component;
+use bevy::{
+    prelude::{Component, ReflectComponent},
+    reflect::Reflect,
+};
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 
@@ -24,9 +27,13 @@ use std::collections::BTreeMap;
 /// what kind of thing they are, e.g. Anchor, Lane, Model, etc. This should be
 /// set by the respective site system that decorates its entities with
 /// components, e.g. add_door_visuals, add_lane_visuals, etc.
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "bevy", derive(Component))]
+#[derive(
+    Serialize, Deserialize, Clone, Copy, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash,
+)]
+#[cfg_attr(feature = "bevy", derive(Component, Reflect))]
+#[cfg_attr(feature = "bevy", reflect(Component))]
 pub enum Category {
+    #[default]
     General,
     Site,
     Anchor,
